requirements.txt
bokeh==3.1.1
Django==4.1.4
djangorestframework==3.14.0
pandas==1.3.5
